Updates to allow for static files and data files in a specific directory; add confirmation to template copy when file exists

+++ packages/parcel-prototyper/src/template.js
@@ -8,6 +8,7 @@ const npmInstall = require('./utils/npmInstall');
 const path = require('path');
 const pathNormalize = require('normalize-path');
 const pkg = require('../package.json');
+const {prompt} = require('enquirer');
 
 class Template {
     /**
@@ -134,10 +135,23 @@ class Template {
                 const templateRelPath = path.normalize(templateFilePath.replace(pathNormalize(templateSrcPath), './'));
                 const projectTemplatePath = path.resolve(entryPath, templateRelPath);
                 const isDirectory = fs.lstatSync(templateFilePath).isDirectory();
+                const exists = fs.existsSync(projectTemplatePath);
 
                 if (isDirectory) {
                     fs.mkdirpSync(projectTemplatePath);
                 } else {
+                    if (exists) {
+                        prompt({
+                            type: 'confirm',
+                            name: 'continue',
+                            message: `Update ${templateRelPath}?`
+                        })
+                        .then(answers => {
+                            if (answers.continue === true) continue;
+                        })
+                        .catch(error => { throw new Error(error) })
+                    }
+                    
                     fs.copySync(templateFilePath, projectTemplatePath);
                 }
             }
